+++ lib/axn/webhooks.rb
@@ -10,6 +10,7 @@ require_relative "webhooks/resolvers"
 require_relative "webhooks/verify"
 require_relative "webhooks/verifiers"
 require_relative "webhooks/verifiers/hmac"
+require_relative "webhooks/verifiers/standard_webhooks"
 require_relative "webhooks/inbound"
 
 module Axn